addde the check for the shade

+++ src/main/java/com/main/glory/servicesImpl/ShadeServiceImpl.java
@@ -192,6 +192,10 @@ public class ShadeServiceImpl implements ShadeServicesInterface {
 		if(shadeMast.isEmpty()){
 			throw new Exception("shade data does not exist with id:"+id);
 		}
+		//check the produciton is avialble or not
+		List<ProductionPlan> productionPlans = productionPlanDao.getAllProductionByShadeId(id);
+		if(!productionPlans.isEmpty())
+			throw new Exception("remove the production first");
 		shadeMastDao.deleteById(id);
 
 		return true;
